AMMI of Genotype by Environment Interaction Model

Performs Additive Main Effects and Multiplication Interaction Analysis of Genotype by Environment Interaction Model

Usage

ge_ammi(.data, .y, .gen, .env, .rep)

# Default S3 method
ge_ammi(.data, .y, .gen, .env, .rep)

Arguments

.data

data.frame

.y

Response Variable

.gen

Genotypes Factor

.env

Environment Factor

.rep

Replication Factor

Value

Genotype by Environment Interaction Model

References

Crossa, J., Perez-Elizalde, S., Jarquin, D., Cotes, J.M., Viele, K., Liu, G., and Cornelius, P.L. (2011) Bayesian Estimation of the Additive Main Effects and Multiplicative Interaction Model Crop Science, 51, 1458–1469. (doi: 10.2135/cropsci2010.06.0343)

Author

Muhammad Yaseen (myaseen208@gmail.com)

Examples

data(Maiz)
fm1 <-
   ge_ammi(
      .data  = Maiz
     , .y    = Yield
     , .gen  = Gen
     , .env  = Env
     , .rep  = Rep
     )
fm1
#> $g
#> [1] 9
#> 
#> $e
#> [1] 20
#> 
#> $Rep
#> [1] 4
#> 
#> $k
#> [1] 8
#> 
#> $ge_means
#> $ge_means$ge_means
#>              1        2        3        4        5        6        7        8
#>  [1,] 3621.550 3728.325 5554.025 4565.575 4379.800 6436.725 2832.400 6011.400
#>  [2,] 3426.450 3918.650 4936.675 4962.575 5201.475 6036.275 2514.600 5277.700
#>  [3,] 3445.850 4082.125 5117.025 5136.275 4177.925 6458.800 3529.075 4730.750
#>  [4,] 3719.900 4539.175 4542.500 6029.500 5672.250 6677.800 2997.600 2516.225
#>  [5,] 3164.625 4078.850 6172.525 5831.000 5413.550 6881.700 3555.600 2731.600
#>  [6,] 4116.075 4877.550 5204.825 5979.900 5591.350 6915.950 3948.875 2982.575
#>  [7,] 3354.025 4767.025 5389.025 4342.275 4277.175 6744.550 3536.875 4205.575
#>  [8,] 4528.775 3393.250 5248.075 4441.500 4476.400 4986.250 3087.675 4483.900
#>  [9,] 3135.525 4500.050 3779.600 5781.375 5407.150 5609.700 3061.375 3309.425
#>              9       10       11       12       13       14       15       16
#>  [1,] 4646.700 3099.875 4432.850 6873.275 6720.550 5848.575 4600.925 5009.675
#>  [2,] 4713.900 2971.750 4348.525 7570.900 5626.975 5932.475 4125.825 5196.450
#>  [3,] 5447.500 2785.375 4526.200 7727.025 6293.875 5886.475 4537.400 5455.475
#>  [4,] 4863.925 2843.325 7117.225 8385.050 7331.725 6439.375 6331.175 6351.100
#>  [5,] 5588.350 2688.000 5995.175 8105.600 7173.575 6358.975 6327.975 6069.800
#>  [6,] 5602.725 3023.500 6150.275 7636.575 7261.925 6380.225 5961.350 5730.175
#>  [7,] 4318.200 2888.925 5051.575 7443.575 5543.850 5819.775 4345.925 5013.425
#>  [8,] 4001.050 3353.125 3712.975 5815.825 4116.550 5522.450 4321.200 4551.000
#>  [9,] 5553.450 2773.950 6429.925 8090.675 6919.875 6282.350 4889.475 5278.375
#>             17       18       19       20
#>  [1,] 4414.850 3344.475 1632.000 4587.025
#>  [2,] 4210.550 4415.200 2281.750 4396.475
#>  [3,] 4749.175 4294.600 3059.175 5018.175
#>  [4,] 5160.850 5617.700 2232.600 4987.975
#>  [5,] 5454.375 4497.800 3072.525 5775.875
#>  [6,] 5807.050 5333.025 3011.025 5088.475
#>  [7,] 3861.575 5276.350 3210.550 4056.075
#>  [8,] 5243.000 2940.375 2634.350 4806.175
#>  [9,] 4988.775 5244.400 2734.950 4822.275
#> 
#> $ge_means$grand_mean
#> [1] 4858.138
#> 
#> 
#> $mu
#> [1] 4858.138
#> 
#> $sigma2
#> [1] 602844.8
#> 
#> $tau
#> [1] 1.658802e-06
#> 
#> $tao
#>            GEffs
#>  [1,] -241.10903
#>  [2,] -254.87903
#>  [3,]  -35.22403
#>  [4,]  359.71097
#>  [5,]  388.73597
#>  [6,]  472.03347
#>  [7,] -185.82153
#>  [8,] -574.94278
#>  [9,]   71.49597
#> 
#> $delta
#> $delta$e_effects
#>             EEffs
#>  [1,] -1245.60722
#>  [2,]  -648.69333
#>  [3,]   246.78167
#>  [4,]   371.85944
#>  [5,]    97.09278
#>  [6,]  1447.16778
#>  [7,] -1628.79611
#>  [8,]  -830.45444
#>  [9,]   112.50667
#> [10,] -1921.71278
#> [11,]   449.05389
#> [12,]  2658.36222
#> [13,]  1473.96222
#> [14,]  1194.15944
#> [15,]   190.89000
#> [16,]   548.02611
#> [17,]    18.55111
#> [18,]  -306.59056
#> [19,] -2206.03500
#> [20,]   -20.52389
#> 
#> 
#> $lambdas
#> [1] 5922.8913 3070.4543 2552.4065 2320.2216 1758.6707 1313.4148  792.6324
#> [8]  757.0277
#> 
#> $alphas
#>             [,1]         [,2]         [,3]        [,4]        [,5]        [,6]
#>  [1,] -0.3812275  0.226376853  0.640917402  0.06276348 -0.27443433  0.37189646
#>  [2,] -0.2475437  0.318181602  0.006538921  0.28188493 -0.01561052 -0.59051506
#>  [3,] -0.1646604  0.148474658  0.057621352 -0.20971703  0.57150726  0.10396503
#>  [4,]  0.5150500 -0.003274222  0.055897289  0.19678126 -0.53669368 -0.15093898
#>  [5,]  0.2830484 -0.342584150  0.367812867 -0.48923056  0.20120360 -0.40537436
#>  [6,]  0.2711150 -0.219555439 -0.091312330 -0.08018815 -0.03226380  0.54258480
#>  [7,] -0.1187550  0.378875355 -0.568803868 -0.51368905 -0.26377705  0.03780326
#>  [8,] -0.4791048 -0.693744264 -0.287949071  0.23087508 -0.08707764 -0.04170396
#>  [9,]  0.3220780  0.187249608 -0.180722562  0.52052004  0.43714616  0.13228281
#>              [,7]        [,8]
#>  [1,] -0.02847968  0.25131927
#>  [2,]  0.46762033 -0.28151218
#>  [3,] -0.45151147 -0.50113650
#>  [4,] -0.44908389 -0.26316390
#>  [5,]  0.11832901  0.31298999
#>  [6,]  0.57650833 -0.35300400
#>  [7,] -0.04118176  0.26668701
#>  [8,] -0.16110446  0.08106783
#>  [9,] -0.03109641  0.48675248
#> 
#> $gammas
#>                [,1]         [,2]         [,3]         [,4]        [,5]
#>  [1,] -0.2247956069 -0.268762762 -0.180394316  0.222348371 -0.30092049
#>  [2,]  0.0175838398  0.177687532 -0.298911720 -0.041784559 -0.09111097
#>  [3,] -0.2576076055 -0.175443927  0.178717786 -0.529519189 -0.18043272
#>  [4,]  0.1479472725 -0.088454266  0.069163882  0.217056712  0.16726564
#>  [5,]  0.0858940079 -0.087735308 -0.015520307  0.332760370 -0.13388850
#>  [6,] -0.0062634265  0.202736159  0.119096124 -0.436497716 -0.22923737
#>  [7,] -0.0811124294 -0.104663893 -0.205425713 -0.255736793  0.15108197
#>  [8,] -0.6729482134  0.417876989  0.208225840  0.270178727 -0.01642642
#>  [9,]  0.0379649453  0.044739257  0.120116710  0.036624299  0.56157302
#> [10,] -0.2181054837 -0.067854487 -0.114378161  0.181480140 -0.15899739
#> [11,]  0.3893229450  0.068645984 -0.059924965  0.161721122 -0.23021612
#> [12,]  0.1599691886  0.315491388  0.024223837 -0.006684231  0.17771568
#> [13,]  0.2494794721  0.243512687  0.502522098 -0.008462921  0.05281381
#> [14,] -0.0066476422  0.023585491 -0.005191355  0.104224942 -0.02038511
#> [15,]  0.2075844883 -0.282722634  0.217172457 -0.120945134 -0.27658890
#> [16,]  0.0789946673  0.002229003  0.112568804 -0.036351911 -0.09773665
#> [17,]  0.0000971901 -0.429651688  0.026360843  0.176686594  0.13223547
#> [18,]  0.2171860426  0.341218990 -0.436647122 -0.014493154 -0.12508018
#> [19,] -0.0881298634 -0.075990205 -0.423893917 -0.250725603  0.37341100
#> [20,] -0.0364137885 -0.256444309  0.162119195 -0.001880067  0.24492422
#>              [,6]         [,7]         [,8]
#>  [1,]  0.22504209 -0.130319195 -0.346996059
#>  [2,]  0.23675510  0.123777955  0.075854726
#>  [3,] -0.24701208  0.236917971  0.284672769
#>  [4,] -0.08520769  0.086797272 -0.253419484
#>  [5,] -0.35236840  0.648776843  0.103439840
#>  [6,]  0.08901469  0.141215583 -0.340199773
#>  [7,]  0.35496791 -0.055017924  0.028131793
#>  [8,]  0.01828035 -0.105003681  0.082344453
#>  [9,]  0.11756614  0.209431533 -0.061495518
#> [10,]  0.01349805 -0.095472591  0.181030943
#> [11,]  0.10857575 -0.305245881  0.533888082
#> [12,] -0.40955254 -0.254230128 -0.098681803
#> [13,]  0.44305168  0.095962075  0.073725166
#> [14,] -0.08463012  0.007199696  0.162724818
#> [15,] -0.02802126 -0.117949268 -0.009900736
#> [16,] -0.28799008 -0.338596778 -0.376137186
#> [17,]  0.20513432  0.072427393 -0.133836476
#> [18,] -0.03944587  0.128626906 -0.181451526
#> [19,] -0.09573202 -0.080505313  0.157417883
#> [20,] -0.18192604 -0.268792467  0.118888087
#>